various aica/arm/misc fixes
set SB_ADST to 1 when dma is in progress. Fixes Carrier audio glitches
(Issue #927)
move aica and rtc scheduler callbacks to aica
implement internal aica dma
slow mmu: missing instruction tlb protection violation. don't flush mmu
tables when changing state
sh4: sr.RB isn't set to 0 in user mode, simply ignored
various non-functional changes

+++ core/hw/sh4/modules/bsc.cpp
@@ -10,7 +10,7 @@ BSC_PDTRA_type BSC_PDTRA;
 
 void write_BSC_PCTRA(u32 addr, u32 data)
 {
-	BSC_PCTRA.full=(u16)data;
+	BSC_PCTRA.full = data;
    if (settings.System == DC_PLATFORM_NAOMI)
       NaomiBoardIDWriteControl((u16)data);
 	//printf("C:BSC_PCTRA = %08X\n",data);
